# perf(fslib): skip unnecessary `realpath` calls in `xfs.mktemp*`